Addition of two brake wear scenarios
Addition of a low brake wear scenario for EVs and high brake wear to account for the variability of brake wear found in the literature. Brake wear for EVs is believed to increase as a result of higher vehicle weight and decrease from the use of regenerative braking.
Two new scenarios model a 20% increase in brake wear and a 80% decrease.